Extract balance accounts table
